compile warnings
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@3633 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/Player.cpp
@@ -8960,13 +8960,13 @@ void Player::EventStunOrImmobilize( Unit* proc_target, bool is_victim )
 
 	if( is_victim == false )
 	{
-		t_trigger_on_stun = trigger_on_stun;
-		t_trigger_on_stun_chance = trigger_on_stun_chance;
+		t_trigger_on_stun = (int32)trigger_on_stun;
+		t_trigger_on_stun_chance = (int32)trigger_on_stun_chance;
 	}
 	else
 	{
-		t_trigger_on_stun = trigger_on_stun_victim;
-		t_trigger_on_stun_chance = trigger_on_stun_chance_victim;
+		t_trigger_on_stun = (int32)trigger_on_stun_victim;
+		t_trigger_on_stun_chance = (int32)trigger_on_stun_chance_victim;
 	}
 
 	if( t_trigger_on_stun )
